Fixes .bashrc being overridden upon startup

+++ termux-shared/src/main/java/com/termux/shared/shell/command/environment/ShellEnvironmentUtils.java
@@ -187,7 +187,9 @@ public class ShellEnvironmentUtils {
                 Logger.logErrorExtended(LOG_TAG, "Failed to create shell home directory\n" + error.toString());
             }
 
-            FileUtils.copyResourceToFile(context, R.raw.bashrc, homeDirectory + "/.bashrc", Charset.defaultCharset());
+            if (!FileUtils.directoryFileExists(homeDirectory + "/.bashrc", true)) {
+                FileUtils.copyResourceToFile(context, R.raw.bashrc, homeDirectory + "/.bashrc", Charset.defaultCharset());
+            }
             setupAppListCache(context, true);
         }
     }
